Implement useContextStateSelector, move context files
