+++ src/context/MapEditor/context.ts
@@ -1,12 +1,5 @@
 import React, { Dispatch } from 'react';
-import {
-  IMapEditorContext,
-  initialState,
-  MapEditorActions,
-  Mode,
-} from './MapEditorReducer';
-
-export { Mode };
+import { IMapEditorContext, initialState, MapEditorActions } from './reducer';
 
 export const MapEditorContext = React.createContext<{
   state: IMapEditorContext;
